requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.5.3
numpy==1.26.4
pandas==2.2.1
scipy==1.11.4
lifelines==0.27.8
pyyaml==6.0.1
jinja2==3.1.3
pyarrow==15.0.0
httpx==0.27.0
pytest==7.4.4
